pip_requirements.txt
addict==2.4.0
aiofiles==24.1.0
aiohttp==3.8.5
aioprocessing==2.0.1
aiosignal==1.3.1
alibabacloud-bailian20230601==1.6.1
alibabacloud-credentials==0.3.5
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.2
alibabacloud-openapi-util==0.2.2
alibabacloud-tea==0.3.9
alibabacloud-tea-openapi==0.3.10
alibabacloud-tea-util==0.3.13
alibabacloud-tea-xml==0.0.2
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
analytics-python==1.2.9
annotated-types==0.6.0
anthropic==0.10.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.3
BitVector==3.5.0
black==24.8.0
bleach==6.1.0
blinker==1.8.2
brotli==1.1.0
burnysc2==6.5.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.13
civrealm==0.1.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.0
crcmod==1.7
cryptography==41.0.5
cssselect2==0.7.0
cycler==0.12.1
dacite==1.8.1
dashscope==1.20.3
dataclasses-json==0.6.7
datasets==2.13.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
diambra==0.0.19
diambra-arena==2.2.7
diambra-engine==2.2.4
dill==0.3.6
distro==1.9.0
dm-env==1.6
dm-haiku==0.0.12
dm-tree==0.1.8
dnspython==2.6.1
docker==7.1.0
docker-pycreds==0.4.0
einops==0.7.0
et-xmlfile==1.1.0
eventlet==0.36.1
exceptiongroup==1.1.3
executing==2.1.0
Farama-Notifications==0.0.4
fastapi==0.103.2
fastjsonschema==2.20.0
filelock==3.12.4
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==4.3.1
flatbuffers==23.5.26
fonttools==4.44.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
gast==0.5.4
gevent==23.7.0
gitdb==4.0.10
gitpython==3.1.32
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.26.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
gql==3.5.0
graphql-core==3.2.4
greenlet==2.0.2
grpcio==1.60.0
grpcio-status
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.2
httptools==0.6.0
httpx==0.27.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
immutabledict==4.2.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
inputs==0.5
ipykernel==6.29.5
ipynbname==2024.1.0.0
ipython==8.18.1
isoduration==20.11.0
itsdangerous==2.2.0
janus==1.0.0
jedi==0.19.1
jinja2==3.1.2
jiter==0.5.0
jmespath==0.10.0
joblib==1.3.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
langchain==0.3.5
langchain-bailian==1.0.0
langchain-community==0.3.3
langchain-core==0.3.13
langchain-text-splitters==0.3.1
langsmith==0.1.137
llvmlite==0.43.0
loguru==0.6.0
marko==2.1.2
markdown-it-py==3.0.0
markdown2==2.4.10
markupsafe==2.1.3
marshmallow==3.21.3
matplotlib==3.8.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml_collections==0.1.1
ml_dtypes==0.5.0
modelscope==1.9.4
monotonic==1.6
mpmath==1.3.0
mpyq==0.2.5
msgpack==1.0.8
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.2
objgraph==3.6.1
ollama==0.3.1
onnxruntime==1.16.0
openai==1.45.0
opencv-python==4.10.0.84
openpyxl==3.1.2
orjson==3.10.6
oss2==2.18.3
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.1
pango==0.0.1
parso==0.8.4
pathspec==0.12.1
pathtools==0.1.2
pillow==10.0.1
pinecone-client==2.2.2
platformdirs==3.11.0
portpicker==1.6.0
posthog==3.0.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.5
pulsar-client==3.3.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.9.2
pydantic-core==2.23.4
pydyf==0.8.0
Pygments==2.18.0
pyjwt==2.8.0
pynndescent==0.5.13
pyparsing==3.1.1
pyphen==0.14.0
pypika==0.48.9
pyreadline3==3.4.1
PySocks==1.7.1
python-box==7.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==3.13.2
python-json-logger==2.0.7
python-socketio==4.6.0
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.13
pyyaml==6.0.1
pyzmq==26.2.0
ray==2.6.3
referencing==0.35.1
regex==2023.10.3
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
roman==4.2
rpds-py==0.20.0
rsa==4.9
s2clientprotocol==5.0.11.90870.0
safetensors==0.4.0
sc2reader==1.8.0
scikit-learn==1.3.1
scipy==1.11.2
screeninfo==0.8.1
seaborn==0.13.0
selenium==4.9.1
Send2Trash==1.8.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.28.1
setproctitle==1.3.2
setuptools==68.0.0
simple-websocket==1.0.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.2.0
tiktoken==0.4.0
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.14.1
tomli==2.0.1
torch==2.1.0
torchvision==0.16.0
tornado==6.3.2
tqdm==4.66.1
traitlets==5.14.3
transformers==4.34.0
treelib==1.7.0
trio==0.26.2
trio-websocket==0.11.1
typer==0.9.0
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2023.3
umap-learn==0.5.6
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.23.2
wandb==0.18.0
watchfiles==0.20.0
wcwidth==0.2.13
weasyprint==60.1
weave==0.50.1
webcolors==24.8.0
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.4
wheel==0.38.4
win32-setctime==1.1.0
wsproto==1.2.0
xlsxwriter==3.1.9
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
zope-event==5.0
zope-interface==6.0
zopfli==0.2.3
antlr4-python3-runtime==4.7.2
lark-parser==0.12.0
multipledispatch==1.0.0
patsy==0.5.6
pddl==0.2.0
pydantic-settings==2.6.0
statsmodels==0.14.4
tarski==0.8.2
ujson==5.10.0
